Refactor and Implementation:
Fixed styling and sizing discrepancies between the `Sources` and `Consumers` pages. Fixed an issue where the tables were overflowing unecessarily causing the page to be wider than it needs to be.

Fixed an issue where the `Create` buttons were not being placed properly by placing them within the same container as the tables.

Added validation on the backend, sending the proper status codes and error messages to the frontend to be displayed to the user.

Fixed TypeErrors in various files and reformatted some code for readability.

- fixed array data when viewed on frontend for topics, endpoints, subscribers
- made client id optional on front end and in user config DB
- cleaned up Kafka endpoint access on backend
- changed fonts across front end
- made buttons consistent in UI
- changed endpoint_url to tumbleweed_endpoint in frontend, backend, and DB
- changed new consumer form (made some things optional, removed others
- fixed alignment and margins on source, consumer, and topic tables on frontend
- Changed header font sizes and spacing on frontend

Co-authored-by: Cruz <102073996+archzedzenrun@users.noreply.github.com>
Co-authored-by: Esther Kim <ekim1009@users.noreply.github.com>
Co-authored-by: Paco Michelson <jeffbbz@users.noreply.github.com>